Added ease of use command to help set various tool and part attributes for durabilities and sharpness. Also fixed Honing breaking tools if the sharpen rate is increased and it happens to skip over the 1% threshold, now it sets it to 1 durability minimum and scales the Sharpness accordingly as well.

+++ Toolsmith/Config/ToolsmithConfigs.cs
@@ -13,7 +13,7 @@ namespace Toolsmith.Config {
         public bool DebugMessages = false;
         public double HeadDurabilityMult = 5.0;
         public double SharpnessMult = 1.5;
-        public int GrindstoneSharpenPerTick = 1;
+        public double GrindstoneSharpenPerTick = 1;
         public double SharpWear = 0.15;
         public double BluntWear = 0.02;
         public float PercentDamageForReforge = 1.0f;
